cascade shadow map 임시 완료
문제는 far 수치를 잘못 넣어주고 있었던 것.
문제 코드는 고쳤으나 magic number 하드코딩해줌.

+++ DUOLGraphicsLibrary_Direct3D11/src/DUOLGraphicsLibrary_Direct3D11/Renderer/RenderTarget/D3D11RenderTarget.cpp
@@ -110,8 +110,8 @@ namespace   DUOLGraphicsLibrary
 	{
 		auto& info = texture->GetTextureDesc();
 
-		_renderTargetDesc._resolution.x = info._textureExtent.x * pow(0.5f, _renderTargetDesc._mipLevel);
-		_renderTargetDesc._resolution.y = info._textureExtent.y * pow(0.5f, _renderTargetDesc._mipLevel);
+		_renderTargetDesc._resolution.x = static_cast<float>(info._textureExtent.x * pow(0.5f, _renderTargetDesc._mipLevel));
+		_renderTargetDesc._resolution.y = static_cast<float>(info._textureExtent.y * pow(0.5f, _renderTargetDesc._mipLevel));
 	}
 
 	void D3D11RenderTarget::CreateRenderTargetViews(ID3D11Device* device, ID3D11Texture2D* texture,
